requirements.txt
airbyte-cdk>=0.80.0
azure-storage-blob>=12.0.0
azure-identity>=1.15.0
